refactor: only one value expression
